fix: Implement user check in the dependency
