+++ agents-api/agents_api/web.py
@@ -6,7 +6,6 @@ import asyncio
 import logging
 from collections.abc import Callable
 from typing import Any, cast
-from uuid import UUID
 
 import sentry_sdk
 import uvicorn
@@ -17,7 +16,6 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
 from litellm.exceptions import APIError
 from pydantic import ValidationError
-from starlette.status import HTTP_403_FORBIDDEN, HTTP_404_NOT_FOUND
 from temporalio.service import RPCError
 
 from .app import app
@@ -25,7 +23,6 @@ from .common.exceptions import BaseCommonException
 from .dependencies.auth import get_api_key
 from .env import enable_responses, sentry_dsn
 from .exceptions import PromptTooBigError
-from .queries.developers.get_developer import get_developer
 from .routers import (
     agents,
     docs,
@@ -282,69 +279,6 @@ else:
     app.include_router(internal.router)
 app.include_router(jobs.router, dependencies=[Depends(get_api_key)])
 app.include_router(healthz.router)
-
-
-# Add middleware for active developer check
-@app.middleware("http")
-async def check_active_developer(request: Request, call_next):
-    """
-    Process the request and check if the developer is active.
-
-    Args:
-        request: The incoming request
-        call_next: The next middleware or endpoint handler
-
-    Returns:
-        The response from the next middleware or endpoint
-    """
-    skip_prefixes = [
-        "/docs",
-        "/redoc",
-        "/openapi.json",
-        "/healthz",
-    ]
-
-    if any(request.url.path.startswith(prefix) for prefix in skip_prefixes):
-        return await call_next(request)
-
-    # Get the X-Developer-Id header
-    developer_id_header = request.headers.get("X-Developer-Id")
-    if not developer_id_header:
-        return await call_next(request)  # Let it fail at the route level if required
-
-    try:
-        # Convert to UUID
-        developer_id = UUID(developer_id_header)
-        # The get_developer function already filters on active=true
-        await get_developer(developer_id=developer_id)
-        # If no exception, developer is active
-    except HTTPException as e:
-        if e.status_code == HTTP_404_NOT_FOUND:
-            # Developer not found or not active
-            return JSONResponse(
-                content={
-                    "error": {
-                        "message": "Invalid developer account",
-                        "code": "invalid_developer_account",
-                    }
-                },
-                status_code=HTTP_403_FORBIDDEN,
-            )
-
-        return JSONResponse(
-            content={
-                "error": {
-                    "message": e.detail,
-                    "code": f"http_{e.status_code}",
-                    "type": "http_error",
-                }
-            },
-            status_code=e.status_code,
-        )
-    except ValueError:
-        return await call_next(request)
-
-
 # TODO: CORS should be enabled only for JWT auth
 #
 app.add_middleware(
